BugFix: 좋아요 조회 쿼리 버그 수정

+++ src/main/java/com/jigumulmi/place/dto/response/ReactionDto.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class ReactionDto {
 
-    private Integer likeReactionCount;
+    private Long likeReactionCount;
     private Long likeReactionId;
 }
